Переписал структуру файлов блокчейна, удалил логирование (будет добавлено в следующей версии).
