+++ core/mine.py
@@ -1,0 +1,56 @@
+import json
+import time
+import requests
+from .block import Block
+from .blockchain_functions import proof_of_work
+from miner_config import MINER_ADDRESS, MINER_NODE_URL
+from .database_functions import get_blockchain_from_file, get_peer_nodes_from_file, write_blockchain_to_memory
+
+
+def mine(a):
+    BLOCKCHAIN = get_blockchain_from_file()
+    NODE_PENDING_TRANSACTIONS = get_peer_nodes_from_file()
+    while True:
+        """Mining is the only way that new coins can be created.
+        In order to prevent too many coins to be created, the process
+        is slowed down by a proof of work algorithm.
+        """
+        # Get the last proof of work
+        last_block = BLOCKCHAIN[-1]
+        last_proof = last_block.data['proof-of-work']
+        # Find the proof of work for the current block being mined
+        # Note: The program will hang here until a new proof of work is found
+        proof = proof_of_work(last_proof)
+        # If we didn't guess the proof, start mining again
+        if not proof[0]:
+            # Update blockchain and save it to file
+            a.send(BLOCKCHAIN)
+            continue
+        else:
+            # Once we find a valid proof of work, we know we can mine a block so
+            # ...we reward the miner by adding a transaction
+            # First we load all pending transactions sent to the node server
+            NODE_PENDING_TRANSACTIONS = requests.get(url=MINER_NODE_URL + '/txion',
+                                                     params={'update': MINER_ADDRESS}).content
+            NODE_PENDING_TRANSACTIONS = json.loads(NODE_PENDING_TRANSACTIONS)
+            # Then we add the mining reward
+            NODE_PENDING_TRANSACTIONS.append({
+                "from": "network",
+                "to": MINER_ADDRESS,
+                "amount": 1})
+            # Now we can gather the data needed to create the new block
+            new_block_data = {
+                "proof-of-work": proof[0],
+                "transactions": list(NODE_PENDING_TRANSACTIONS)
+            }
+            new_block_index = last_block.index + 1
+            new_block_timestamp = time.time()
+            last_block_hash = last_block.hash
+            # Empty transaction list
+            NODE_PENDING_TRANSACTIONS = []
+            # Now create the new block
+            mined_block = Block(new_block_index, new_block_timestamp, new_block_data, last_block_hash)
+            BLOCKCHAIN.append(mined_block)
+            write_blockchain_to_memory(BLOCKCHAIN)
+            a.send(BLOCKCHAIN)
+            requests.get(url=MINER_NODE_URL + '/blocks', params={'update': MINER_ADDRESS})
